revert: lifecycle-integration sprint (PR #511 + hotfixes #512 #513 #514)
Rolls back the 10-PR integration and its 3 subsequent migration hotfixes.
Prod baseline restored. Migration 0025 kept in source because it is
already applied to prod D1 (the meetings/meeting_schedule tables exist
and are harmless — additive, unused by pre-integration code). Migration
0026 removed because every variant we tried failed at the DROP TABLE
entities step.

## What we know about 0026

Three speculative fixes all failed remote D1 execution:
  1. UPDATE-before-shadow-swap → CHECK constraint (expected)
  2. PRAGMA defer_foreign_keys = ON → D1 internal error (reproducibly)
  3. PRAGMA foreign_keys = off → FK constraint (pragma did not persist
     across statements in D1's migrate-apply path)

Hypothesis: wrangler's migrations-apply splits SQL by `;` and runs each
statement in its own transaction, so pragmas set at the top do not
persist to the DROP TABLE later. This differs from upstream SQLite
and from the pattern suggested by D1's own docs example — needs
investigation against a clone of prod D1, not more speculative pushes.

## Follow-up needed (separate issue)

- Reproduce the failure against a fresh D1 seeded with a small
  'assessing' row + an FK-dependent row; identify whether the fix
  requires combined-statement SQL, `d1 execute` batching, or a
  fundamentally different migration strategy.
- Once 0026 has been proven on a staging D1, re-apply the integration.

Prod state summary:
- Worker: reverts to pre-integration code (what users have been seeing)
- D1: 0025 applied (harmless), 0026 not applied
- Source: matches prod code; 0025 present so wrangler sanity check passes
