--- .playwright-mcp/console-2026-04-23T21-36-46-158Z.log
@@ -1,20 +1,0 @@
-[     245ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/@vite/client:0
-[     245ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/@id/astro/runtime/client/dev-toolbar/entrypoint.js:0
-[     246ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/src/styles/global.css:0
-[     416ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/favicon.svg:0
-[   36578ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/@vite/client:0
-[   36578ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/@id/astro/runtime/client/dev-toolbar/entrypoint.js:0
-[   36578ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/src/styles/global.css:0
-[   52631ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/@vite/client:0
-[   52632ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/@id/astro/runtime/client/dev-toolbar/entrypoint.js:0
-[   52632ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/src/styles/global.css:0
-[ 1068968ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/@vite/client:0
-[ 1068968ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/@id/astro/runtime/client/dev-toolbar/entrypoint.js:0
-[ 1068968ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/src/styles/global.css:0
-[ 1068968ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/src/pages/design-preview/modern-institutional/index.astro?astro&type=style&index=0&lang.css:0
-[ 1071019ms] [ERROR] Failed to load resource: the server responded with a status of 500 (Internal Server Error) @ http://localhost:4321/api/events:0
-[ 1561890ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/@id/astro/runtime/client/dev-toolbar/entrypoint.js:0
-[ 1561890ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/@vite/client:0
-[ 1561890ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/src/styles/global.css:0
-[ 1561891ms] [ERROR] Failed to load resource: the server responded with a status of 404 (Not Found) @ http://localhost:4321/src/pages/design-preview/modern-institutional/index.astro?astro&type=style&index=0&lang.css:0
-[ 1563941ms] [ERROR] Failed to load resource: the server responded with a status of 500 (Internal Server Error) @ http://localhost:4321/api/events:0
