--- .playwright-mcp/console-2026-04-23T22-20-15-110Z.log
@@ -1,3 +1,0 @@
-[    3082ms] [INFO] %cDownload the React DevTools for a better development experience: https://reactjs.org/link/react-devtools font-weight:bold @ https://unpkg.com/react-dom@18.3.1/umd/react-dom.development.js:29904
-[    3262ms] [WARNING] You are using the in-browser Babel transformer. Be sure to precompile your scripts for production - https://babeljs.io/docs/setup/ @ https://unpkg.com/@babel/standalone@7.29.0/babel.min.js:2
-[    3915ms] [ERROR] Failed to load resource: the server responded with a status of 404 () @ https://09b5a8c7-5c15-4c84-af3f-76d8dac6fe8e.claudeusercontent.com/v1/design/projects/09b5a8c7-5c15-4c84-af3f-76d8dac6fe8e/serve/ui_kits/comparison/.design-canvas.state.json:0
