--- src/components/admin/LogReplyDialog.astro
@@ -1,158 +1,0 @@
----
-/**
- * LogReplyDialog — modal for capturing a prospect reply (issue #464).
- *
- * Emits a POST to /api/admin/entities/[id]/reply-log which creates a context
- * entry of type `note` with source `reply_log` and structured metadata.
- *
- * Uses a native <dialog> element so it works without a JS framework.
- * Opened by any button with `data-open-reply-dialog="<entityId>"`.
- */
-
-export interface Props {
-  entityId: string
-  /** Display name — used in the dialog header so admin knows which entity this is for. */
-  entityName: string
-  /** Unique suffix so multiple dialogs can coexist on one page (list view). */
-  instanceId?: string
-}
-
-const { entityId, entityName, instanceId = entityId } = Astro.props
-const dialogId = `reply-dialog-${instanceId}`
-const formId = `reply-form-${instanceId}`
----
-
-<dialog
-  id={dialogId}
-  data-reply-dialog
-  data-entity-id={entityId}
-  class="rounded-[var(--radius-card)] border border-[color:var(--color-border)] p-0 backdrop:bg-black/40 w-full max-w-md"
->
-  <form
-    id={formId}
-    method="POST"
-    action={`/api/admin/entities/${entityId}/reply-log`}
-    class="bg-white p-card"
-  >
-    <div class="flex items-start justify-between mb-stack">
-      <div>
-        <h3 class="text-base font-semibold text-[color:var(--color-text-primary)]">Log reply</h3>
-        <p class="text-xs text-[color:var(--color-text-muted)] mt-0.5">
-          {entityName}
-        </p>
-      </div>
-      <button
-        type="button"
-        data-close-dialog
-        class="text-xs text-[color:var(--color-text-muted)] hover:text-[color:var(--color-text-primary)] transition-colors"
-        aria-label="Close"
-      >
-        Close
-      </button>
-    </div>
-
-    <div class="space-y-stack">
-      <div>
-        <label
-          class="block text-xs font-medium text-[color:var(--color-text-secondary)] mb-1"
-          for={`${instanceId}-sentiment`}
-        >
-          Sentiment
-        </label>
-        <select
-          id={`${instanceId}-sentiment`}
-          name="sentiment"
-          required
-          class="w-full text-sm border border-[color:var(--color-border)] rounded-[var(--radius-card)] px-3 py-2 bg-white text-[color:var(--color-text-primary)] focus:outline-none focus:ring-2 focus:ring-primary/20 focus:border-primary"
-        >
-          <option value="interested">Interested</option>
-          <option value="declined">Declined</option>
-          <option value="out_of_office">Out of office</option>
-          <option value="other">Other</option>
-        </select>
-      </div>
-
-      <div>
-        <label
-          class="block text-xs font-medium text-[color:var(--color-text-secondary)] mb-1"
-          for={`${instanceId}-next-action`}
-        >
-          Suggested next action
-        </label>
-        <select
-          id={`${instanceId}-next-action`}
-          name="next_action"
-          required
-          class="w-full text-sm border border-[color:var(--color-border)] rounded-[var(--radius-card)] px-3 py-2 bg-white text-[color:var(--color-text-primary)] focus:outline-none focus:ring-2 focus:ring-primary/20 focus:border-primary"
-        >
-          <option value="book_meeting">Book meeting</option>
-          <option value="retry_later">Retry later</option>
-          <option value="mark_lost">Mark lost</option>
-          <option value="continue_conversation">Continue conversation</option>
-        </select>
-        <p class="text-xs text-[color:var(--color-text-muted)] mt-1">
-          No stage change happens automatically. Use the stage buttons on the entity after saving.
-        </p>
-      </div>
-
-      <div>
-        <label
-          class="block text-xs font-medium text-[color:var(--color-text-secondary)] mb-1"
-          for={`${instanceId}-notes`}
-        >
-          Notes
-        </label>
-        <textarea
-          id={`${instanceId}-notes`}
-          name="notes"
-          rows="3"
-          placeholder="Paste or summarize the reply..."
-          class="w-full text-sm border border-[color:var(--color-border)] rounded-[var(--radius-card)] px-3 py-2 resize-y focus:outline-none focus:ring-2 focus:ring-primary/20 focus:border-primary"
-        ></textarea>
-      </div>
-    </div>
-
-    <div class="mt-card flex items-center justify-end gap-2">
-      <button
-        type="button"
-        data-close-dialog
-        class="text-sm bg-[color:var(--color-background)] text-[color:var(--color-text-secondary)] px-3 py-1.5 rounded-[var(--radius-card)] hover:bg-[color:var(--color-border-subtle)] transition-colors"
-      >
-        Cancel
-      </button>
-      <button
-        type="submit"
-        class="text-sm bg-primary text-white px-4 py-1.5 rounded-[var(--radius-card)] hover:bg-primary/90 transition-colors"
-      >
-        Save reply
-      </button>
-    </div>
-  </form>
-</dialog>
-
-<script>
-  // Wire up open/close for every reply dialog on the page. Each trigger uses
-  // data-open-reply-dialog="<entityId>" to target its dialog.
-  const openers = document.querySelectorAll<HTMLElement>('[data-open-reply-dialog]')
-  openers.forEach((el) => {
-    el.addEventListener('click', (e) => {
-      const entityId = el.getAttribute('data-open-reply-dialog')
-      if (!entityId) return
-      // Match by data-entity-id so multiple dialogs on the same page pick the right one.
-      const dialog = document.querySelector<HTMLDialogElement>(
-        `dialog[data-reply-dialog][data-entity-id="${entityId}"]`
-      )
-      if (dialog) {
-        e.preventDefault()
-        dialog.showModal()
-      }
-    })
-  })
-
-  document.querySelectorAll<HTMLElement>('[data-close-dialog]').forEach((btn) => {
-    btn.addEventListener('click', () => {
-      const dialog = btn.closest('dialog') as HTMLDialogElement | null
-      dialog?.close()
-    })
-  })
-</script>
