+++ src/lib/db/entities.ts
@@ -13,7 +13,6 @@
 import { computeSlug } from '../entities/slug.js'
 import { recomputeDeterministicCache } from '../entities/recompute.js'
 import { appendContext } from './context.js'
-import { isLostReasonCode, type LostReasonCode } from './lost-reasons.js'
 
 // ---------------------------------------------------------------------------
 // Types
@@ -44,7 +43,7 @@ export interface Entity {
 export type EntityStage =
   | 'signal'
   | 'prospect'
-  | 'meetings'
+  | 'assessing'
   | 'proposing'
   | 'engaged'
   | 'delivered'
@@ -64,7 +63,7 @@ export type EntityVertical =
 export const ENTITY_STAGES: { value: EntityStage; label: string }[] = [
   { value: 'signal', label: 'Signal' },
   { value: 'prospect', label: 'Prospect' },
-  { value: 'meetings', label: 'Meetings' },
+  { value: 'assessing', label: 'Assessing' },
   { value: 'proposing', label: 'Proposing' },
   { value: 'engaged', label: 'Engaged' },
   { value: 'delivered', label: 'Delivered' },
@@ -94,14 +93,8 @@ export const ENTITY_VERTICALS: { value: EntityVertical; label: string }[] = [
  */
 const VALID_TRANSITIONS: Record<EntityStage, EntityStage[]> = {
   signal: ['prospect', 'lost'],
-  prospect: ['meetings', 'lost'],
-  // From `meetings` the admin picks the next step explicitly (#470). Direct
-  // transitions to `engaged`/`delivered`/`ongoing` still require going
-  // through `proposing` first — the `proposing→engaged` accepted-quote
-  // invariant protects the engagement model. Backing out to `prospect` is
-  // allowed so a discovery/follow-up meeting that didn't qualify doesn't
-  // force an entity into `lost`.
-  meetings: ['proposing', 'prospect', 'lost'],
+  prospect: ['assessing', 'lost'],
+  assessing: ['proposing', 'lost'],
   proposing: ['engaged', 'lost'],
   engaged: ['delivered', 'lost'],
   delivered: ['ongoing', 'prospect', 'lost'],
@@ -143,21 +136,6 @@ export type FindOrCreateResult =
 export interface TransitionStageOptions {
   /** Override reason — bypasses pre-condition checks where documented. */
   force?: string
-  /**
-   * Structured metadata for `lost` transitions. Captured on the
-   * `stage_change` context entry's JSON metadata so the Lost tab can
-   * filter and future reporting can roll up "why we lost" without
-   * parsing free text.
-   *
-   * Required when `newStage === 'lost'`. Enforced at the DAL layer
-   * rather than the API so every caller (admin UI, scripts, future
-   * background jobs) is held to the same contract.
-   */
-  lostReason?: {
-    code: LostReasonCode
-    /** Optional operator note. Trimmed. Empty → stored as null. */
-    detail?: string | null
-  }
 }
 
 // ---------------------------------------------------------------------------
@@ -248,125 +226,6 @@ export async function countEntitiesByStage(
     .bind(orgId, stage)
     .first<{ count: number }>()
   return result?.count ?? 0
-}
-
-// ---------------------------------------------------------------------------
-// Signal metadata (for Signal list evidence density)
-// ---------------------------------------------------------------------------
-
-/**
- * Per-entity rollup of pipeline-generated signal metadata + last activity,
- * used to render evidence-dense signal rows without loading full context.
- *
- * Values come from the context table: `top_problems` and `outreach_angle`
- * are read from the metadata JSON of the most recent `signal` / `scorecard`
- * context entry; `last_activity_at` is the `created_at` of the most recent
- * context entry of ANY type.
- *
- * Missing fields stay `null` — callers must render nothing (not placeholders)
- * per CLAUDE.md Pattern B.
- */
-export interface EntitySignalMetadata {
-  entity_id: string
-  top_problems: string[] | null
-  outreach_angle: string | null
-  last_activity_at: string | null
-}
-
-/**
- * Fetch latest signal metadata and last-activity timestamp for a batch of
- * entities in two parameterized queries (no N+1).
- *
- * Returns a Map keyed by entity_id. Entities with no context entries at all
- * are omitted from the map — caller should treat missing as "no metadata".
- */
-export async function getSignalMetadataForEntities(
-  db: D1Database,
-  orgId: string,
-  entityIds: string[]
-): Promise<Map<string, EntitySignalMetadata>> {
-  const out = new Map<string, EntitySignalMetadata>()
-  if (entityIds.length === 0) return out
-
-  const placeholders = entityIds.map(() => '?').join(', ')
-
-  // Latest signal/scorecard metadata per entity.
-  // Picks the most recent row via the correlated subquery on created_at.
-  const signalSql = `
-    SELECT c.entity_id, c.metadata
-    FROM context c
-    WHERE c.org_id = ?
-      AND c.entity_id IN (${placeholders})
-      AND c.type IN ('signal', 'scorecard')
-      AND c.created_at = (
-        SELECT MAX(c2.created_at)
-        FROM context c2
-        WHERE c2.entity_id = c.entity_id
-          AND c2.type IN ('signal', 'scorecard')
-      )
-  `
-  const signalRows = await db
-    .prepare(signalSql)
-    .bind(orgId, ...entityIds)
-    .all<{ entity_id: string; metadata: string | null }>()
-
-  for (const row of signalRows.results) {
-    let topProblems: string[] | null = null
-    let outreachAngle: string | null = null
-    if (row.metadata) {
-      try {
-        const meta = JSON.parse(row.metadata) as Record<string, unknown>
-        if (
-          Array.isArray(meta.top_problems) &&
-          meta.top_problems.every((p) => typeof p === 'string')
-        ) {
-          topProblems = (meta.top_problems as string[]).length
-            ? (meta.top_problems as string[])
-            : null
-        }
-        if (typeof meta.outreach_angle === 'string' && meta.outreach_angle.trim()) {
-          outreachAngle = meta.outreach_angle.trim()
-        }
-      } catch {
-        // Malformed JSON — treat as missing metadata.
-      }
-    }
-    out.set(row.entity_id, {
-      entity_id: row.entity_id,
-      top_problems: topProblems,
-      outreach_angle: outreachAngle,
-      last_activity_at: null,
-    })
-  }
-
-  // Last-activity across all context types.
-  const activitySql = `
-    SELECT entity_id, MAX(created_at) AS last_activity_at
-    FROM context
-    WHERE org_id = ?
-      AND entity_id IN (${placeholders})
-    GROUP BY entity_id
-  `
-  const activityRows = await db
-    .prepare(activitySql)
-    .bind(orgId, ...entityIds)
-    .all<{ entity_id: string; last_activity_at: string | null }>()
-
-  for (const row of activityRows.results) {
-    const existing = out.get(row.entity_id)
-    if (existing) {
-      existing.last_activity_at = row.last_activity_at
-    } else {
-      out.set(row.entity_id, {
-        entity_id: row.entity_id,
-        top_problems: null,
-        outreach_angle: null,
-        last_activity_at: row.last_activity_at,
-      })
-    }
-  }
-
-  return out
 }
 
 // ---------------------------------------------------------------------------
@@ -547,8 +406,8 @@ export async function updateEntity(
  * - proposing → engaged: requires at least one accepted quote
  * - delivered → ongoing: requires paid completion invoice OR force override
  *
- * Note: signal → meetings is blocked by VALID_TRANSITIONS. Booking flows
- * must walk through `prospect` as an intermediate state (signal → prospect → meetings).
+ * Note: signal → assessing is blocked by VALID_TRANSITIONS. Booking flows
+ * must walk through `prospect` as an intermediate state (signal → prospect → assessing).
  *
  * Records a stage_change context entry automatically.
  */
@@ -568,28 +427,6 @@ export async function transitionStage(
     throw new Error(
       `Invalid stage transition: ${entity.stage} → ${newStage}. Allowed: ${allowed?.join(', ')}`
     )
-  }
-
-  // Lost reason is required when transitioning to lost. Captured as
-  // structured metadata on the stage_change context entry so the Lost
-  // tab can filter and "why we lost" rollups are queryable.
-  let lostReasonCode: LostReasonCode | null = null
-  let lostReasonDetail: string | null = null
-  if (newStage === 'lost') {
-    if (!options?.lostReason?.code) {
-      throw new Error(
-        'Lost reason is required: provide options.lostReason.code when transitioning to lost.'
-      )
-    }
-    if (!isLostReasonCode(options.lostReason.code)) {
-      throw new Error(
-        `Invalid lost reason code: ${options.lostReason.code}. See src/lib/db/lost-reasons.ts.`
-      )
-    }
-    lostReasonCode = options.lostReason.code
-    const rawDetail = options.lostReason.detail
-    lostReasonDetail =
-      typeof rawDetail === 'string' && rawDetail.trim().length > 0 ? rawDetail.trim() : null
   }
 
   // ---------------------------------------------------------------------------
@@ -653,80 +490,26 @@ export async function transitionStage(
   // Record stage change as context entry
   const contextId = crypto.randomUUID()
   const content = `Stage: ${entity.stage} → ${newStage}. ${reason}`
-  const metadata: Record<string, unknown> = {
-    from: entity.stage,
-    to: newStage,
-    reason,
-  }
-  if (lostReasonCode) {
-    metadata.lost_reason = lostReasonCode
-    if (lostReasonDetail) metadata.lost_detail = lostReasonDetail
-  }
   await db
     .prepare(
       `INSERT INTO context (id, entity_id, org_id, type, content, source, content_size, metadata, created_at)
       VALUES (?, ?, ?, 'stage_change', ?, 'system', ?, ?, ?)`
     )
-    .bind(contextId, entityId, orgId, content, content.length, JSON.stringify(metadata), now)
+    .bind(
+      contextId,
+      entityId,
+      orgId,
+      content,
+      content.length,
+      JSON.stringify({ from: entity.stage, to: newStage, reason }),
+      now
+    )
     .run()
 
   // Recompute cache after stage change
   await recomputeDeterministicCache(db, orgId, entityId)
 
   return getEntity(db, orgId, entityId)
-}
-
-/**
- * Returns the latest captured Lost reason code per entity, keyed by
- * entity_id. Reads from `stage_change` context entries where
- * `metadata.to = 'lost'`. Entities with no structured reason (e.g.
- * legacy Lost rows captured before #477) are absent from the map.
- *
- * This is the one place that rolls up structured Lost metadata for
- * list-rendering. Keep the query tight — it runs on every Lost-tab
- * page render.
- */
-export async function getLatestLostReasonsByEntity(
-  db: D1Database,
-  orgId: string,
-  entityIds: string[]
-): Promise<Map<string, { code: LostReasonCode; detail: string | null }>> {
-  const result = new Map<string, { code: LostReasonCode; detail: string | null }>()
-  if (entityIds.length === 0) return result
-
-  const placeholders = entityIds.map(() => '?').join(', ')
-  // For each entity, pick the newest stage_change row whose metadata
-  // marks a transition INTO lost. D1/SQLite supports json_extract.
-  const rows = await db
-    .prepare(
-      `SELECT c.entity_id,
-              json_extract(c.metadata, '$.lost_reason') AS lost_reason,
-              json_extract(c.metadata, '$.lost_detail') AS lost_detail
-         FROM context c
-        WHERE c.org_id = ?
-          AND c.type = 'stage_change'
-          AND c.entity_id IN (${placeholders})
-          AND json_extract(c.metadata, '$.to') = 'lost'
-          AND c.created_at = (
-            SELECT MAX(c2.created_at) FROM context c2
-             WHERE c2.org_id = c.org_id
-               AND c2.entity_id = c.entity_id
-               AND c2.type = 'stage_change'
-               AND json_extract(c2.metadata, '$.to') = 'lost'
-          )`
-    )
-    .bind(orgId, ...entityIds)
-    .all<{ entity_id: string; lost_reason: string | null; lost_detail: string | null }>()
-
-  for (const row of rows.results) {
-    if (row.lost_reason && isLostReasonCode(row.lost_reason)) {
-      result.set(row.entity_id, {
-        code: row.lost_reason,
-        detail: row.lost_detail ?? null,
-      })
-    }
-  }
-  return result
 }
 
 // ---------------------------------------------------------------------------
