+++ src/lib/db/quotes.ts
@@ -17,20 +17,7 @@ export interface Quote {
   id: string
   org_id: string
   entity_id: string
-  /**
-   * Legacy FK into assessments.id. NOT NULL at the schema level. During the
-   * monitoring window this value equals meeting_id (#469 preserves IDs across
-   * the meetings backfill) and assessments rows are dual-written on intake.
-   * A follow-up drop migration removes both the column and the assessments
-   * table together.
-   */
   assessment_id: string
-  /**
-   * New reference to meetings.id (#469). Nullable for pre-migration rows
-   * until the backfill in migration 0025 populates it from assessment_id.
-   * New code should prefer this column.
-   */
-  meeting_id: string | null
   version: number
   parent_quote_id: string | null
   line_items: string // JSON array of LineItem
@@ -110,19 +97,7 @@ export const VALID_TRANSITIONS: Record<QuoteStatus, QuoteStatus[]> = {
 
 export interface CreateQuoteData {
   entityId: string
-  /**
-   * Legacy assessment id. Still written to the NOT NULL `assessment_id`
-   * column until the follow-up drop migration removes it. Callers should
-   * pass `meetingId` for new work; when both are present they must match
-   * (and currently do by construction — see migration 0025).
-   */
   assessmentId: string
-  /**
-   * New canonical reference to meetings.id (#469). When omitted the caller
-   * is asserting this quote pre-dates the meetings migration; in that case
-   * createQuote copies assessmentId into meeting_id for forward compatibility.
-   */
-  meetingId?: string
   lineItems: LineItem[]
   rate: number
   depositPct?: number
@@ -130,13 +105,6 @@ export interface CreateQuoteData {
   deliverables?: DeliverableRow[]
   engagementOverview?: string
   milestoneLabel?: string
-  /**
-   * Optional reference to a prior quote this one supersedes. Used by the
-   * repeat-quote flow (#472) to link v2 → v1 after a decline/expiry. The
-   * caller is responsible for explicitly transitioning the parent to
-   * `superseded` — createQuote does not mutate the parent record.
-   */
-  parentQuoteId?: string | null
 }
 
 export interface UpdateQuoteData {
@@ -210,33 +178,6 @@ export function getMissingAuthoredContent(quote: Quote): string[] {
 }
 
 /**
- * Quote statuses that block a new draft from being created on the same entity.
- * Draft and sent are "open" — actively in play in the sales flow. Other
- * statuses (accepted, declined, expired, superseded) are terminal and do not
- * block a repeat quote (#472).
- */
-export const OPEN_QUOTE_STATUSES: QuoteStatus[] = ['draft', 'sent']
-
-/**
- * Return true if the entity has at least one draft or sent quote.
- * Used by the repeat-quote flow (#472) to gate the "New quote" action on the
- * entity detail page — admins shouldn't be creating v2 while v1 is still live.
- */
-export async function hasOpenQuoteForEntity(
-  db: D1Database,
-  orgId: string,
-  entityId: string
-): Promise<boolean> {
-  const placeholders = OPEN_QUOTE_STATUSES.map(() => '?').join(', ')
-  const sql = `SELECT 1 FROM quotes WHERE entity_id = ? AND org_id = ? AND status IN (${placeholders}) LIMIT 1`
-  const result = await db
-    .prepare(sql)
-    .bind(entityId, orgId, ...OPEN_QUOTE_STATUSES)
-    .first<{ '1': number }>()
-  return result !== null
-}
-
-/**
  * List quotes for an organization, optionally filtered by entity.
  */
 export async function listQuotes(
@@ -303,39 +244,17 @@ export async function createQuote(
     data.deliverables && data.deliverables.length > 0 ? JSON.stringify(data.deliverables) : null
   const engagementOverview = data.engagementOverview ?? null
   const milestoneLabel = data.milestoneLabel ?? null
-  const parentQuoteId = data.parentQuoteId ?? null
-
-  // Derive the next version number when superseding a prior quote so the
-  // portal's "latest version" lookup (parent_quote_id OR assessment_id + version)
-  // still resolves correctly. Standalone new quotes start at version 1.
-  let version = 1
-  if (parentQuoteId) {
-    const parent = await db
-      .prepare('SELECT version FROM quotes WHERE id = ? AND org_id = ?')
-      .bind(parentQuoteId, orgId)
-      .first<{ version: number }>()
-    if (parent) {
-      version = (parent.version ?? 1) + 1
-    }
-  }
-
-  // meeting_id mirrors assessment_id by construction (meetings preserve the
-  // assessment primary key) unless the caller passes an explicit meetingId.
-  const meetingIdValue = data.meetingId ?? data.assessmentId
 
   await db
     .prepare(
-      `INSERT INTO quotes (id, org_id, entity_id, assessment_id, meeting_id, version, parent_quote_id, line_items, total_hours, rate, total_price, deposit_pct, deposit_amount, status, schedule, deliverables, engagement_overview, milestone_label, created_at, updated_at)
-     VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, 'draft', ?, ?, ?, ?, ?, ?)`
+      `INSERT INTO quotes (id, org_id, entity_id, assessment_id, version, line_items, total_hours, rate, total_price, deposit_pct, deposit_amount, status, schedule, deliverables, engagement_overview, milestone_label, created_at, updated_at)
+     VALUES (?, ?, ?, ?, 1, ?, ?, ?, ?, ?, ?, 'draft', ?, ?, ?, ?, ?, ?)`
     )
     .bind(
       id,
       orgId,
       data.entityId,
       data.assessmentId,
-      meetingIdValue,
-      version,
-      parentQuoteId,
       JSON.stringify(data.lineItems),
       totalHours,
       data.rate,
@@ -487,71 +406,6 @@ export async function updateQuote(
  * Draft and superseded quotes are internal-only.
  */
 const PORTAL_VISIBLE_STATUSES = ['sent', 'accepted', 'declined', 'expired'] as const
-
-/**
- * For a set of entity ids, return the "top active" quote per entity — the one
- * an admin scanning a list would care about. Priority:
- *
- *   1. Most recently sent `sent` quote (the one in the client's hands)
- *   2. Otherwise most recently updated `accepted` quote
- *   3. Otherwise most recently updated `draft` quote
- *
- * Terminal non-useful statuses (declined, expired, superseded) are ignored.
- * Returns a Map keyed by entity_id so the caller can render badges inline on
- * a list without an N+1 per-row query.
- *
- * Scoped by org_id for tenant isolation. Empty id list short-circuits to an
- * empty map (D1 rejects `IN ()`).
- */
-export async function getActiveQuotesForEntities(
-  db: D1Database,
-  orgId: string,
-  entityIds: string[]
-): Promise<Map<string, Quote>> {
-  const result = new Map<string, Quote>()
-  if (entityIds.length === 0) return result
-
-  const placeholders = entityIds.map(() => '?').join(', ')
-  // Priority: sent (0) > accepted (1) > draft (2). We order by priority then by
-  // `sent_at` DESC (for sent), else `updated_at` DESC. One row per entity via
-  // the ROW_NUMBER() window function. D1 supports SQLite window functions.
-  const sql = `
-    SELECT id, org_id, entity_id, assessment_id, version, parent_quote_id,
-           line_items, total_hours, rate, total_price, deposit_pct,
-           deposit_amount, status, sent_at, expires_at, accepted_at,
-           schedule, deliverables, engagement_overview, milestone_label,
-           created_at, updated_at
-    FROM (
-      SELECT q.*,
-        ROW_NUMBER() OVER (
-          PARTITION BY q.entity_id
-          ORDER BY
-            CASE q.status
-              WHEN 'sent'     THEN 0
-              WHEN 'accepted' THEN 1
-              WHEN 'draft'    THEN 2
-              ELSE 9
-            END,
-            COALESCE(q.sent_at, q.updated_at) DESC
-        ) AS rn
-      FROM quotes q
-      WHERE q.org_id = ?
-        AND q.entity_id IN (${placeholders})
-        AND q.status IN ('sent', 'accepted', 'draft')
-    )
-    WHERE rn = 1
-  `
-
-  const rows = await db
-    .prepare(sql)
-    .bind(orgId, ...entityIds)
-    .all<Quote>()
-
-  for (const row of rows.results) {
-    result.set(row.entity_id, row)
-  }
-  return result
-}
 
 /**
  * List quotes for a specific entity (portal access).
