+++ src/pages/admin/analytics/index.astro
@@ -118,7 +118,7 @@ function formatShortDate(iso: string): string {
 // ── Pipeline helpers ─────────────────────────────────────────────
 const pipelineStages = [
   { key: 'prospect' as const, label: 'Prospect', color: 'bg-slate-400' },
-  { key: 'meetings' as const, label: 'Meetings', color: 'bg-blue-500' },
+  { key: 'assessing' as const, label: 'Assessing', color: 'bg-blue-500' },
   { key: 'proposing' as const, label: 'Proposing', color: 'bg-amber-500' },
   { key: 'engaged' as const, label: 'Engaged', color: 'bg-green-500' },
   { key: 'delivered' as const, label: 'Delivered', color: 'bg-emerald-600' },
