+++ src/pages/admin/index.astro
@@ -82,7 +82,7 @@ const signalCount = signals.length
 const pipelineStages = [
   { label: 'Signals', count: signalCount, stage: 'signal', color: 'amber' },
   { label: 'Prospects', count: pipeline.prospect, stage: 'prospect', color: 'blue' },
-  { label: 'Meetings', count: pipeline.meetings, stage: 'meetings', color: 'indigo' },
+  { label: 'Assessing', count: pipeline.assessing, stage: 'assessing', color: 'indigo' },
   { label: 'Proposing', count: pipeline.proposing, stage: 'proposing', color: 'purple' },
   { label: 'Engaged', count: pipeline.engaged, stage: 'engaged', color: 'green' },
   { label: 'Delivered', count: pipeline.delivered, stage: 'delivered', color: 'teal' },
