+++ src/pages/api/admin/assessments/[id]/complete.ts
@@ -14,23 +14,10 @@ import { env } from 'cloudflare:workers'
 const DEFAULT_RATE = 175
 
 /**
- * POST /api/admin/assessments/:id/complete — legacy, kept for backward compat
+ * POST /api/admin/assessments/:id/complete
  *
- * The Meetings generalization (#468, #470) split this endpoint into two new
- * routes scoped under the entity:
- *   - `/api/admin/entities/:id/meetings/:meetingId/complete` — captures
- *     outcome + explicit next-stage picker, no quote side-effects
- *   - `/api/admin/entities/:id/meetings/:meetingId/draft-quote` — explicit
- *     opt-in proposal drafting
- *
- * This legacy handler is unreachable from the admin UI (the
- * `/admin/assessments/[id]` page now 301-redirects to the meeting URL) but
- * stays in place for direct callers / external scripts during the monitoring
- * window. A follow-up issue tracks removing it alongside the assessments
- * table drop.
- *
- * Behavior below is unchanged: builds extraction JSON, marks completed,
- * drafts a quote, and transitions to proposing.
+ * Completes an assessment: builds extraction JSON, transitions status,
+ * appends context, creates a draft quote, and redirects to the new quote.
  *
  * Protected by auth middleware (requires admin role).
  */
@@ -134,11 +121,7 @@ export const POST: APIRoute = async ({ request, locals, redirect, params }) => {
       source_ref: assessmentId,
     })
 
-    // If disqualified, transition to lost and redirect back.
-    // Disqualification during assessment maps to the `not-a-fit` lost
-    // reason code — see src/lib/db/lost-reasons.ts. The extraction's
-    // free-text explanation is carried in `lost_detail` so the admin
-    // can still read the LLM's reasoning when reviewing the Lost tab.
+    // If disqualified, transition to lost and redirect back
     if (disqualified) {
       try {
         await updateAssessmentStatus(env.DB, session.orgId, assessmentId, 'disqualified')
@@ -147,13 +130,7 @@ export const POST: APIRoute = async ({ request, locals, redirect, params }) => {
           session.orgId,
           entity.id,
           'lost',
-          `Disqualified during assessment: ${extraction.disqualify_reason ?? 'No reason provided'}`,
-          {
-            lostReason: {
-              code: 'not-a-fit',
-              detail: extraction.disqualify_reason ?? null,
-            },
-          }
+          `Disqualified during assessment: ${extraction.disqualify_reason ?? 'No reason provided'}`
         )
       } catch {
         // Stage transition may fail if already in lost state
@@ -180,12 +157,10 @@ export const POST: APIRoute = async ({ request, locals, redirect, params }) => {
       lineItems = []
     }
 
-    // 7. Create draft quote with pre-filled line items. meetingId is equal
-    //    to assessmentId by construction (#469 preserved IDs during backfill).
+    // 7. Create draft quote with pre-filled line items
     const quote = await createQuote(env.DB, session.orgId, {
       entityId: entity.id,
       assessmentId,
-      meetingId: assessmentId,
       lineItems,
       rate: DEFAULT_RATE,
     })
