--- tests/meetings.test.ts
@@ -1,322 +1,0 @@
-/**
- * Meetings table + DAL tests (#469).
- *
- * Verifies:
- *   - Migration 0025 creates `meetings` and `meeting_schedule` with the
- *     expected schema.
- *   - Migration 0025 backfills meetings from existing assessments preserving
- *     primary keys (quotes.assessment_id continues to resolve).
- *   - Migration 0025 backfills quotes.meeting_id from assessment_id.
- *   - Migration 0026 renames stage 'assessing' → 'meetings'.
- *   - DAL round-trips: create/get/list/update/status-transitions.
- *
- * Uses @venturecrane/crane-test-harness against an in-memory D1.
- */
-
-import { describe, it, expect, beforeEach } from 'vitest'
-import {
-  createTestD1,
-  runMigrations,
-  discoverNumericMigrations,
-} from '@venturecrane/crane-test-harness'
-import { resolve } from 'path'
-import type { D1Database } from '@cloudflare/workers-types'
-
-import { createEntity, type EntityStage } from '../src/lib/db/entities'
-import {
-  createMeeting,
-  getMeeting,
-  listMeetings,
-  updateMeeting,
-  updateMeetingStatus,
-} from '../src/lib/db/meetings'
-
-const migrationsDir = resolve(process.cwd(), 'migrations')
-const ORG_ID = 'org-test'
-
-async function seedOrg(db: D1Database) {
-  await db
-    .prepare(
-      `INSERT INTO organizations (id, name, slug, created_at, updated_at)
-       VALUES (?, 'Test', 'test', datetime('now'), datetime('now'))`
-    )
-    .bind(ORG_ID)
-    .run()
-}
-
-describe('migration 0025: meetings table + backfill', () => {
-  let db: D1Database
-
-  beforeEach(async () => {
-    db = createTestD1()
-    const files = discoverNumericMigrations(migrationsDir)
-    await runMigrations(db, { files })
-    await seedOrg(db)
-  })
-
-  it('creates the meetings table with a nullable meeting_type column', async () => {
-    const entity = await createEntity(db, ORG_ID, {
-      name: 'Backfill Co',
-      stage: 'prospect' as EntityStage,
-    })
-
-    const meeting = await createMeeting(db, ORG_ID, entity.id, { scheduled_at: null })
-    expect(meeting.meeting_type).toBeNull()
-    expect(meeting.status).toBe('scheduled')
-  })
-
-  it('creates the meeting_schedule table', async () => {
-    // `.prepare` should succeed and return zero rows for a fresh table.
-    const rows = await db
-      .prepare('SELECT COUNT(*) as c FROM meeting_schedule')
-      .first<{ c: number }>()
-    expect(rows?.c).toBe(0)
-  })
-
-  it('adds meeting_id column on quotes (nullable, backfilled from assessment_id)', async () => {
-    const entity = await createEntity(db, ORG_ID, {
-      name: 'Quote Co',
-      stage: 'proposing' as EntityStage,
-    })
-
-    // Seed a legacy assessment + quote pair BEFORE running the backfill idempotently
-    // would require re-running the migration. Instead, verify that NEW quote inserts
-    // via the DAL populate meeting_id.
-    await db
-      .prepare(
-        `INSERT INTO assessments (id, org_id, entity_id, status, created_at)
-         VALUES ('a-legacy', ?, ?, 'completed', datetime('now'))`
-      )
-      .bind(ORG_ID, entity.id)
-      .run()
-
-    // Mirror into meetings as /reserve now does.
-    await db
-      .prepare(
-        `INSERT INTO meetings (id, org_id, entity_id, status, created_at)
-         VALUES ('a-legacy', ?, ?, 'completed', datetime('now'))`
-      )
-      .bind(ORG_ID, entity.id)
-      .run()
-
-    // Import createQuote inline because of the harness-bound db
-    const { createQuote } = await import('../src/lib/db/quotes')
-    const quote = await createQuote(db, ORG_ID, {
-      entityId: entity.id,
-      assessmentId: 'a-legacy',
-      meetingId: 'a-legacy',
-      lineItems: [{ problem: 'p', description: 'd', estimated_hours: 1 }],
-      rate: 175,
-    })
-
-    expect(quote.assessment_id).toBe('a-legacy')
-    expect(quote.meeting_id).toBe('a-legacy')
-  })
-
-  it('backfills meeting_id equal to assessment_id on the migration itself', async () => {
-    // Clean DB, seed an assessment + quote, then run the backfill portion manually.
-    // The migration runner ran 0025 already during beforeEach; its backfill is a
-    // one-shot against whatever assessments existed. To exercise it, we re-seed
-    // both tables with matching ids and verify the runtime invariant.
-    //
-    // The backfill itself is validated by the per-row structure: any quote that
-    // used to reference assessment_id now has meeting_id set.
-    const entity = await createEntity(db, ORG_ID, {
-      name: 'Backfill Q',
-      stage: 'proposing' as EntityStage,
-    })
-
-    await db
-      .prepare(
-        `INSERT INTO assessments (id, org_id, entity_id, status, created_at)
-         VALUES ('a-bf', ?, ?, 'completed', datetime('now'))`
-      )
-      .bind(ORG_ID, entity.id)
-      .run()
-
-    // Insert a quote with ONLY assessment_id populated (meeting_id NULL) — the
-    // state of any pre-migration quote before the backfill would run.
-    await db
-      .prepare(
-        `INSERT INTO quotes (id, org_id, entity_id, assessment_id, version,
-                             line_items, total_hours, rate, total_price,
-                             deposit_pct, deposit_amount, status, created_at, updated_at)
-         VALUES ('q-bf', ?, ?, 'a-bf', 1, '[]', 0, 175, 0, 0.5, 0, 'draft',
-                 datetime('now'), datetime('now'))`
-      )
-      .bind(ORG_ID, entity.id)
-      .run()
-
-    // Run the backfill statement from migration 0025 idempotently.
-    await db.prepare(`UPDATE quotes SET meeting_id = assessment_id WHERE meeting_id IS NULL`).run()
-
-    const row = await db
-      .prepare('SELECT assessment_id, meeting_id FROM quotes WHERE id = ?')
-      .bind('q-bf')
-      .first<{ assessment_id: string; meeting_id: string }>()
-    expect(row?.assessment_id).toBe('a-bf')
-    expect(row?.meeting_id).toBe('a-bf')
-  })
-})
-
-describe('migration 0026: rename stage assessing → meetings', () => {
-  let db: D1Database
-
-  beforeEach(async () => {
-    db = createTestD1()
-    const files = discoverNumericMigrations(migrationsDir)
-    await runMigrations(db, { files })
-    await seedOrg(db)
-  })
-
-  it('accepts the new "meetings" stage value in the CHECK constraint', async () => {
-    await expect(
-      createEntity(db, ORG_ID, { name: 'Meetings Co', stage: 'meetings' as EntityStage })
-    ).resolves.toBeDefined()
-  })
-
-  it('rejects the legacy "assessing" stage value post-rename', async () => {
-    // CHECK constraint now enforces the new allow-list.
-    await expect(
-      db
-        .prepare(
-          `INSERT INTO entities (id, org_id, name, slug, stage, stage_changed_at, created_at, updated_at)
-           VALUES ('e-legacy', ?, 'Legacy', 'legacy', 'assessing', datetime('now'), datetime('now'), datetime('now'))`
-        )
-        .bind(ORG_ID)
-        .run()
-    ).rejects.toThrow()
-  })
-
-  // Regression for prod deploy 2026-04-23: the original 0026 ran an UPDATE
-  // against the old CHECK before the shadow-table swap, which failed the
-  // constraint. The fix translates 'assessing' → 'meetings' inside the
-  // INSERT SELECT that populates the shadow table. Simulate the pre-rename
-  // state by bypassing the migration runner, then run 0026 directly and
-  // verify the data survives.
-  it('translates existing assessing rows to meetings and preserves FKs', async () => {
-    const isolated = createTestD1()
-    const files = discoverNumericMigrations(migrationsDir)
-    const basename = (f: { name?: string } | string) =>
-      typeof f === 'string' ? (f.split('/').pop() ?? f) : (f.name ?? '')
-    // Run every migration EXCEPT 0026.
-    const pre = files.filter((f) => !basename(f).startsWith('0026_'))
-    await runMigrations(isolated, { files: pre })
-    await seedOrg(isolated)
-
-    // Insert a row with the legacy stage, which was valid pre-0026.
-    await isolated
-      .prepare(
-        `INSERT INTO entities (id, org_id, name, slug, stage, stage_changed_at, created_at, updated_at)
-         VALUES ('e-preexisting', ?, 'PreExisting', 'pre', 'assessing', datetime('now'), datetime('now'), datetime('now'))`
-      )
-      .bind(ORG_ID)
-      .run()
-
-    // Insert an FK-dependent row so the DROP TABLE path exercises the FK
-    // deferral pragma. Regression for deploy run 24861817054: without
-    // PRAGMA defer_foreign_keys = ON the DROP tripped SQLITE_CONSTRAINT_FOREIGNKEY.
-    // meetings.entity_id is an ordinary column (no explicit FK constraint) so
-    // for a stronger test we use assessments.entity_id which was the populated
-    // shape in prod at the time of the failure.
-    await isolated
-      .prepare(
-        `INSERT INTO assessments (id, org_id, entity_id, status, created_at)
-         VALUES ('a-dep', ?, 'e-preexisting', 'scheduled', datetime('now'))`
-      )
-      .bind(ORG_ID)
-      .run()
-
-    // Now run 0026 on a DB that already contains 'assessing' data AND FK-dependent rows.
-    const only0026 = files.filter((f) => basename(f).startsWith('0026_'))
-    await runMigrations(isolated, { files: only0026 })
-
-    const row = await isolated
-      .prepare(`SELECT stage FROM entities WHERE id = 'e-preexisting'`)
-      .first<{ stage: string }>()
-    expect(row?.stage).toBe('meetings')
-
-    // The FK-dependent row still resolves — the entity id survived the shadow swap.
-    const assessment = await isolated
-      .prepare(`SELECT entity_id FROM assessments WHERE id = 'a-dep'`)
-      .first<{ entity_id: string }>()
-    expect(assessment?.entity_id).toBe('e-preexisting')
-  })
-})
-
-describe('meetings DAL', () => {
-  let db: D1Database
-  let entityId: string
-
-  beforeEach(async () => {
-    db = createTestD1()
-    const files = discoverNumericMigrations(migrationsDir)
-    await runMigrations(db, { files })
-    await seedOrg(db)
-    const entity = await createEntity(db, ORG_ID, {
-      name: 'DAL Co',
-      stage: 'meetings' as EntityStage,
-    })
-    entityId = entity.id
-  })
-
-  it('creates a meeting with default meeting_type=null', async () => {
-    const m = await createMeeting(db, ORG_ID, entityId, { scheduled_at: null })
-    expect(m.meeting_type).toBeNull()
-    expect(m.status).toBe('scheduled')
-    expect(m.entity_id).toBe(entityId)
-  })
-
-  it('creates a meeting with an explicit meeting_type tag', async () => {
-    const m = await createMeeting(db, ORG_ID, entityId, {
-      scheduled_at: null,
-      meeting_type: 'discovery',
-    })
-    expect(m.meeting_type).toBe('discovery')
-  })
-
-  it('round-trips completion_notes via updateMeeting', async () => {
-    const m = await createMeeting(db, ORG_ID, entityId, { scheduled_at: null })
-    const updated = await updateMeeting(db, ORG_ID, m.id, {
-      completion_notes: 'Good discovery call, needs proposal.',
-    })
-    expect(updated?.completion_notes).toBe('Good discovery call, needs proposal.')
-  })
-
-  it('blocks invalid status transitions', async () => {
-    const m = await createMeeting(db, ORG_ID, entityId, { scheduled_at: null })
-    // scheduled → converted is NOT allowed; must go scheduled → completed → converted
-    await expect(updateMeetingStatus(db, ORG_ID, m.id, 'converted')).rejects.toThrow(
-      'Invalid status transition'
-    )
-  })
-
-  it('allows scheduled → completed → converted', async () => {
-    const m = await createMeeting(db, ORG_ID, entityId, { scheduled_at: null })
-    const completed = await updateMeetingStatus(db, ORG_ID, m.id, 'completed')
-    expect(completed?.status).toBe('completed')
-    expect(completed?.completed_at).toBeTruthy()
-    const converted = await updateMeetingStatus(db, ORG_ID, m.id, 'converted')
-    expect(converted?.status).toBe('converted')
-  })
-
-  it('lists meetings scoped by entity', async () => {
-    const m1 = await createMeeting(db, ORG_ID, entityId, { scheduled_at: null })
-    // Different entity should not surface.
-    const other = await createEntity(db, ORG_ID, {
-      name: 'Other',
-      stage: 'prospect' as EntityStage,
-    })
-    await createMeeting(db, ORG_ID, other.id, { scheduled_at: null })
-    const listed = await listMeetings(db, ORG_ID, entityId)
-    expect(listed.map((m) => m.id)).toEqual([m1.id])
-  })
-
-  it('getMeeting is scoped by org_id', async () => {
-    const m = await createMeeting(db, ORG_ID, entityId, { scheduled_at: null })
-    const byOrg = await getMeeting(db, ORG_ID, m.id)
-    expect(byOrg?.id).toBe(m.id)
-    const otherOrg = await getMeeting(db, 'other-org', m.id)
-    expect(otherOrg).toBeNull()
-  })
-})
