+++ workers/job-monitor/src/index.ts
@@ -15,7 +15,6 @@ import { findOrCreateEntity } from '../../../src/lib/db/entities.js'
 import { appendContext } from '../../../src/lib/db/context.js'
 import { getGeneratorConfig, recordGeneratorRun } from '../../../src/lib/db/generators.js'
 import type { JobMonitorConfig } from '../../../src/lib/generators/types.js'
-import { enrichEntity } from '../../../src/lib/enrichment/index.js'
 import { searchJobs } from './serpapi.js'
 import { qualifyJob, derivePainScore } from './qualify.js'
 import { sendFailureAlert, type RunSummary } from './alert.js'
@@ -27,13 +26,9 @@ export interface Env {
   ANTHROPIC_API_KEY: string
   RESEND_API_KEY: string
   LEAD_INGEST_API_KEY: string
-  // Optional API keys consumed by the at-ingest enrichment pipeline.
-  GOOGLE_PLACES_API_KEY?: string
-  OUTSCRAPER_API_KEY?: string
-  PROXYCURL_API_KEY?: string
 }
 
-async function run(env: Env, ctx?: ExecutionContext): Promise<RunSummary> {
+async function run(env: Env): Promise<RunSummary> {
   const summary: RunSummary = {
     queries: 0,
     totalResults: 0,
@@ -164,16 +159,6 @@ async function run(env: Env, ctx?: ExecutionContext): Promise<RunSummary> {
       })
 
       summary.written++
-
-      // At-ingest enrichment (issue #471). Detached so the ingest loop is not
-      // blocked by N Claude round-trips. Idempotent — no-op if a prior brief
-      // exists on this entity already.
-      const enrichPromise = enrichEntity(env, ORG_ID, entity.id, { mode: 'full' }).catch((err) => {
-        console.error('[job_monitor] enrichment failed for', entity.id, err)
-      })
-      if (ctx) {
-        ctx.waitUntil(enrichPromise)
-      }
     } catch (err) {
       summary.errors++
       const msg = err instanceof Error ? err.message : String(err)
@@ -197,7 +182,7 @@ async function run(env: Env, ctx?: ExecutionContext): Promise<RunSummary> {
 
 export default {
   async scheduled(controller: ScheduledController, env: Env, ctx: ExecutionContext): Promise<void> {
-    const summary = await run(env, ctx)
+    const summary = await run(env)
     if (summary.written === 0 && summary.errors > 0 && env.RESEND_API_KEY) {
       ctx.waitUntil(sendFailureAlert(summary, env.RESEND_API_KEY))
     }
@@ -208,7 +193,7 @@ export default {
     if (auth !== `Bearer ${env.LEAD_INGEST_API_KEY}`) {
       return new Response('Unauthorized', { status: 401 })
     }
-    const summary = await run(env, ctx)
+    const summary = await run(env)
     return new Response(JSON.stringify(summary, null, 2), {
       headers: { 'Content-Type': 'application/json' },
     })
